+++ src/Factories/CacheFactory.php
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-namespace GrahamCampbell\Throttle\Classes;
+namespace GrahamCampbell\Throttle\Factories;
 
 use Illuminate\Http\Request;
 use Illuminate\Cache\Repository;
+use GrahamCampbell\Throttle\Throttlers\CacheThrottler;
 
 /**
- * This is the throttle class.
+ * This is the cache throttler factory class.
  *
  * @package    Laravel-Throttle
  * @author     Graham Campbell
@@ -28,7 +29,7 @@ use Illuminate\Cache\Repository;
  * @license    https://github.com/GrahamCampbell/Laravel-Throttle/blob/master/LICENSE.md
  * @link       https://github.com/GrahamCampbell/Laravel-Throttle
  */
-class Throttle
+class CacheFactory implements FactoryInterface
 {
     /**
      * The cache instance.
@@ -38,68 +39,31 @@ class Throttle
     protected $cache;
 
     /**
-     * The throttler.
-     *
-     * @var string
-     */
-    protected $throttler;
-
-    /**
      * Create a new instance.
      *
      * @param  \Illuminate\Cache\Repository  $cache
-     * @param  string  $throttler
      * @return void
      */
-    public function __construct(Repository $cache, $throttler)
+    public function __construct(Repository $cache)
     {
         $this->cache = $cache;
-        $this->throttler = $throttler;
     }
 
     /**
-     * Get the throttler.
+     * Make a new throttler instance.
      *
      * @param  array|\Illuminate\Http\Request  $data
      * @param  int  $limit
      * @param  int  $time
      * @return \GrahamCampbell\Throttle\Throttlers\ThrottlerInterface
      */
-    public function get($data, $limit = 10, $time = 60)
+    public function make($data, $limit = 10, $time = 60)
     {
         $data = $this->parseData($data);
         $store = $this->getStore($data['ip']);
         $key = $this->getKey($data['route']);
 
-        $throttler = $this->throttler;
-
-        return new $throttler($store, $key, $limit, $time);
-    }
-
-    /**
-     * Hit the the throttle.
-     *
-     * @param  array|\Illuminate\Http\Request  $data
-     * @param  int  $limit
-     * @param  int  $time
-     * @return \GrahamCampbell\Throttle\Throttlers\ThrottlerInterface
-     */
-    public function hit($data, $limit = 10, $time = 60)
-    {
-        return $this->get($data, $limit, $time)->hit();
-    }
-
-    /**
-     * Clear the the throttle.
-     *
-     * @param  array|\Illuminate\Http\Request  $data
-     * @param  int  $limit
-     * @param  int  $time
-     * @return \GrahamCampbell\Throttle\Throttlers\ThrottlerInterface
-     */
-    public function clear($data, $limit = 10, $time = 60)
-    {
-        return $this->get($data, $limit, $time)->clear();
+        return new CacheThrottler($store, $key, $limit, $time);
     }
 
     /**