+++ tests/ServiceProviderTest.php
@@ -30,4 +30,14 @@ use GrahamCampbell\TestBench\Traits\ServiceProviderTestCaseTrait;
 class ServiceProviderTest extends AbstractTestCase
 {
     use ServiceProviderTestCaseTrait;
+
+    public function testThrottleFactoryIsInjectable()
+    {
+        $this->assertIsInjectable('GrahamCampbell\Throttle\Factories\FactoryInterface');
+    }
+
+    public function testThrottleIsInjectable()
+    {
+        $this->assertIsInjectable('GrahamCampbell\Throttle\Throttle');
+    }
 }