+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGE LOG
 
 ## V1.0 Beta 1 (Upcoming - July 2014)
 
-* Upgrade to Laravel 4.2
+* Upgrade to Laravel 4.1+
 
 
 ## V0.2 Alpha (21/04/2014)